+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] 2019-08-24
 ### Added
 - Json parsing with HttpResponse method
 - More Timeouts options
@@ -59,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - https
 
 
-[Unreleased]: https://github.com/sonic182/aiosonic/compare/0.3.1..HEAD
+[Unreleased]: https://github.com/sonic182/aiosonic/compare/0.4.0..HEAD
+[0.4.0]: https://github.com/sonic182/aiosonic/compare/0.3.1..0.4.0
 [0.3.1]: https://github.com/sonic182/aiosonic/compare/0.3.0..0.3.1
 [0.3.0]: https://github.com/sonic182/aiosonic/compare/0.2.1..0.3.0
 [0.2.1]: https://github.com/sonic182/aiosonic/compare/0.1.0..0.2.1